+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to
 - Fixed bug where typing decimals into spinboxes wasn't working
 - Fixed bug where sky was shown a second time outside field of view with
   equidistant and equal area camera projections
+- Fixed missing prism face distances from save files
 
 ## 3.2.0 - 2021-03-24